% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_deploy.R
\name{append_task_to_fabfile}
\alias{append_task_to_fabfile}
\title{append_task_to_fabfile}
\usage{
append_task_to_fabfile(entry, conf, job_name)
}
\description{
Appends task function definition to fabfile.
}
