% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_deploy.R
\name{build_parallel_run_job}
\alias{build_parallel_run_job}
\title{build_parallel_run_job}
\usage{
build_parallel_run_job(conf, job_name, args, iter_arg = NULL)
}
\description{
Builds jenkins job for parallel execution of task.
}
