% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classified_names.R
\docType{data}
\name{classified_names}
\alias{classified_names}
\title{classified_names}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35311 rows and 2 columns.}
\usage{
classified_names
}
\description{
A dataframe derived from ssa_national that contains names, classified
by the historical liklihood of gender membership, from 1978 to 1989.
}
\keyword{datasets}
