% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_bind.R
\name{clean_bind}
\alias{clean_bind}
\title{clean_bind}
\usage{
clean_bind(frames)
}
\arguments{
\item{frames}{A list of data frames to bind}
}
\description{
A safer way to invoke bind_rows that fill not fail if dataframes are null.
}
