% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_conform_funs.R
\name{clean_country}
\alias{clean_country}
\title{clean_country}
\usage{
clean_country(country, name_type = "country_name", return_original = FALSE)
}
\arguments{
\item{country}{A vector of unconformed state values.}

\item{name_type}{One of country_name, continent, or region.  Defaults to country_name.}

\item{return_original}{Should the input value be returned if a match cannot be found?
Defaults to FALSE.  If no match found, NA is returned.}
}
\description{
Conforms country names by matching against a number of internatinoal standards.
Specifically: is02c, iso3c, FIPS, cocw, and IOC.  Additionally, can return continent and region.
}
