% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_conform_funs.R
\name{clean_gpa}
\alias{clean_gpa}
\title{clean_gpa}
\usage{
clean_gpa(x)
}
\arguments{
\item{x}{A vector of unconformed gpa values.}
}
\description{
Converts gpa range values to numeric.
Returns NA is value is 0, or greater than 4.5.
}
