% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_conform_funs.R
\name{clean_level_of_education}
\alias{clean_level_of_education}
\title{clean_level_of_education}
\usage{
clean_level_of_education(x, fill.missing = NA)
}
\arguments{
\item{x}{A vector of unconformed education level values.}

\item{fill.missing}{Replace missing values with string (default NA)}
}
\description{
Converts messy education level data to cleaned education levels.
}
