% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_conform_funs.R
\name{clean_military_affiliated}
\alias{clean_military_affiliated}
\title{clean_military_affiliated}
\usage{
clean_military_affiliated(x)
}
\arguments{
\item{x}{A vector of unconformed military affiliation values.}
}
\description{
Converts military affilation values into boolean.
}
