% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_conform_funs.R
\name{clean_state}
\alias{clean_state}
\title{clean_state}
\usage{
clean_state(state, no_match_val = "unmatched", name_type = "abbrev")
}
\arguments{
\item{state}{A vector of unconformed state values.}

\item{no_match_val}{The value to be returned if input value cannot be matched.
Defaults to 'unmatched'}

\item{name_type}{Type of state value to be returned.
One of: abbrev, division, region, or full_name.
Defaults to abbrev.}
}
\description{
Converts messy state data to abbreviated state names.
}
