% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_conform_funs.R
\name{clean_work_exp}
\alias{clean_work_exp}
\title{clean_work_exp}
\usage{
clean_work_exp(x)
}
\arguments{
\item{x}{A vector of unconformed gpa values.}
}
\description{
Converts work experience values to numeric, returning NA for any
value greater than 50, and maximum of 20 for any value below that.
}
