% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbConn.R
\name{conn_bigquery}
\alias{conn_bigquery}
\title{conn_bigquery}
\usage{
conn_bigquery(creds, cred_location, dataset)
}
\arguments{
\item{creds}{Path to a service account token with additional keys for project_id, and dataset.}
}
\description{
DBI interface for bigquery.
}
