% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbConn.R
\name{conn_mssql}
\alias{conn_mssql}
\title{conn_msql}
\usage{
conn_mssql(creds)
}
\arguments{
\item{creds}{A list with entries for db_name, host, user, password, and port.}
}
\description{
DBI interface for microsoft sql server.
}
