% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\name{conn_mysql}
\alias{conn_mysql}
\title{conn_mysql}
\usage{
conn_mysql(creds)
}
\arguments{
\item{creds}{A list with entries for db_name, host, user, password, and port.}
}
\description{
DBI interface for mysql.
}
