% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\name{conn_psql}
\alias{conn_psql}
\title{conn_psql}
\usage{
conn_psql(creds)
}
\arguments{
\item{creds}{A list with entries for db_name, host, user, password, and port.}
}
\description{
DBI interface for postgresql.
}
