% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\name{conn_redshift}
\alias{conn_redshift}
\title{conn_redshift}
\usage{
conn_redshift(creds)
}
\arguments{
\item{creds}{A list with entries for db_name, host, user, password, and port.}
}
\description{
DBI interface for amazon redshift.
}
