% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\name{conn_rpostgres}
\alias{conn_rpostgres}
\title{conn_rpsql}
\usage{
conn_rpostgres(creds)
}
\arguments{
\item{creds}{A list with entries for db_name, host, user, password, and port.}
}
\description{
DBI interface for PostgreSQL using the RPostgres library.
}
