% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsql.R
\name{countsql}
\alias{countsql}
\title{countsql}
\usage{
countsql(col, table, aggCol = 1, aggFunc = "sum", conn = NULL)
}
\arguments{
\item{col}{Character. The column(s) to group by}

\item{table}{Character. The table to query (Schema optional)}

\item{aggCol}{Character. Optional. If left out, returns number of rows in group, similar to n(),
otherwise, returns sum of aggCol.}

\item{aggFunc}{Aggregation function to be applied to data. Defaults to sum.}

\item{conn}{Optional. dplyr connection object.}
}
\description{
SQL Convenience function. Passes vars to SQL GROUP BY verb.
Direct complaints to Mark Thompson.
}
