% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\docType{class}
\name{db_conn-class}
\alias{db_conn-class}
\alias{db_conn}
\title{send}
\arguments{
\item{database}{Name of database for which the connection should created.  Must correspond to a creds.json file key.}

\item{cred_location}{Path to a json credentials file with keys corresponding to database names.
Each database entry should have the following fields: db_name, user, password, host, port
Defaults to creds.json in user root directory.}

\item{database_type}{One of postgres, mysql or ms_sql (for fdw connections). Defaults to postgres}

\item{init_sql}{A sql statement to execute on connection; optional.}

\item{table}{Name of table for which columns should be returned.}

\item{match_string}{String to search for in columns.
Partial matching applied. Optional.}

\item{schema}{Name of schema for which tables should be returned. If blank,
all tables will be returned.}

\item{match_string}{String to search for in columns.
Partial matching applied. Optional.}
}
\description{
Execute arbitrary sql, expecting no response.
Str_form substitution params supported by default.

Return a list of columns from the requested table.

Return a list of tables from the requested schema.
}
\details{
Delivering dead simple database connections since June 7th 2016.
}
\section{Fields}{

\describe{
\item{\code{connection}}{A dplyr src connection object used to execute queries.}

\item{\code{con}}{A DBI compliant connection, extracted from dplyr_src.}
}}

\section{Methods}{

\describe{
\item{\code{get_cols(table, match_string = NA)}}{Return a list of columns in the specified table.
Can optionally apply a regex match to find a specific column.}

\item{\code{query(statement, ...)}}{Execute query using initialized database connection. Str_form templating is supported by default.}

\item{\code{send(statement, ...)}}{Send a statement using initialized database connection. Str_form templating is supported by default.}
}}

