% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\name{db_conn}
\alias{db_conn}
\title{db_con}
\usage{
db_conn(
  database,
  ...,
  cred_location = "~/creds.json",
  database_type = "postgres",
  init_sql = NA,
  dataset = NA
)
}
\arguments{
\item{database}{Name of database for which the connection should created.  Must correspond to a creds.json file key.}

\item{cred_location}{Path to a json credentials file with keys corresponding to database names.
Each database entry should have the following fields: db_name, user, password, host, port
Defaults to creds.json in user root directory.}

\item{init_sql}{A sql statement to execute on connection; optional.}
}
\description{
Delivering dead simple database connections since June 7th 2016.
}
