% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbConn.R
\docType{class}
\name{DbConn}
\alias{DbConn}
\title{read_sql}
\format{\code{\link{R6Class}} object.}
\usage{
DbConn
}
\arguments{
\item{database}{Name of database for which the connection should created.  Must correspond to a creds.json file key.}

\item{cred_location}{Path to a json credentials file with keys corresponding to database names.
Each database entry should have the following fields: db_name, user, password, host, port, database_type
Defaults to creds.json in user root directory.}

\item{init_sql}{A sql statement to execute on connection; optional.}

\item{table}{Name of table for which columns should be returned.}

\item{match_string}{String to search for in columns.
Partial matching applied. Optional.}

\item{schema}{Name of schema for which tables should be returned. If blank,
all tables will be returned.}

\item{match_string}{String to search for in columns.
Partial matching applied. Optional.}
}
\value{
Object of \code{\link{R6Class}} with methods for query execution.
}
\description{
Delivering dead simple database connections since June 7th 2016.

Read a sql statement from disk.

Execute arbitrary sql, expecting a dataframe in response.
Str_form substitution params supported by default.

Disconnect from current database.

Execute arbitrary sql, expecting no response.
Str_form substitution params supported by default.

Return a list of columns from the requested table.

Return a list of tables from the requested schema.
}
\details{
DbConn
}
\section{Fields}{

\describe{
\item{\code{connection}}{A dplyr src connection object used to execute queries.}

\item{\code{con}}{A DBI compliant connection, extracted from dplyr_src.}
}}

\keyword{datasets}
