% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_deploy.R
\name{deploy_project}
\alias{deploy_project}
\title{deploy_project}
\usage{
deploy_project(iter_arg = NULL, deploy_conf = "jenkins/deploy.yaml",
  packrat = TRUE)
}
\arguments{
\item{iter_arg}{Name of argument over which parallel execution job should be bult.}

\item{deploy_file}{Path to yaml file specifying arguments and function calls.}
}
\description{
Writes fabric deployment info based on specifications in yaml file.
Should always be called from the root project directory.
}
