% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid.R
\name{expand_grid}
\alias{expand_grid}
\title{expand_grid}
\usage{
expand_grid(...)
}
\arguments{
\item{...}{Arguments to be passed to expand.grid.}
}
\description{
Does the same thing as expand.grid, just without the insane requirement
to specify stringsAsFactors = FALSE.
}
