% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_deploy.R
\name{export_jobs_to_jenkins}
\alias{export_jobs_to_jenkins}
\title{export_jobs_to_jenkins}
\usage{

  export_jobs_to_jenkins(user = yaml::yaml.load_file("jenkins/deploy.yaml")$user_email,
  api_token = Sys.getenv("jenkins_api_token"),
  project = yaml::yaml.load_file("jenkins/deploy.yaml")$project_name,
  host = "https://corp-bi-jenkins-prod.2u.com")
}
\arguments{
\item{user}{Jenkins user email address used to authenticate. Defaults to
project user email found in jenkins deploy.yaml file.}

\item{api_token}{The api token associated with the jenkins user being used.
By default, mmkit will look for a system enviroment var called jenkins_api_token.}

\item{host}{Hostname of the jenkins server where jobs should be created.
Defaults to https://corp-bi-jenkins-prod.2u.com.  Don't change this.}

\item{project_name}{Project name - Used as prefix for jobs being created.
By default, mmkit will try to read this information from the jenkins/deploy.yaml file.}
}
\description{
Exports the pre-populated xml files in the jenkins directory to
the server.
}
