% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_conform_funs.R
\name{extract_numeric}
\alias{extract_numeric}
\title{extract_numeric}
\usage{
extract_numeric(x, max_value = Inf, na_cap = NULL)
}
\arguments{
\item{x}{A vector of unconformed values.}

\item{max_value}{The maximum value that should be returned. Defaults to infinity.
eg: for gpa values, the cap should be 4.0.}

\item{na_cap}{A numeric above which NA sbould be returned. Defaults to null. Should
be used to invalidate values that don't make sense for the context; eg: a gpa value = 75.}
}
\description{
Converts character value to numeric by removing character and or
punctuation, splitting and averaging if the value is a range (eg: 2.5-3.0 = 2.75),
and imposing a cap on value size.  Also allows the setting of a limit above which NA
should be returned.  Is intended to be used for conforming gpa, work experience,
and other values where numbers are represented as strings that may or may not be a range.
}
