% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConn.R
\name{findConn}
\alias{findConn}
\title{findConn}
\usage{
findConn(env = .GlobalEnv)
}
\arguments{
\item{env}{The environment to search in. Defaults to .GlobalEnv}
}
\description{
Finds and returns an active dplyr Redshift connection object
}
