% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_not_na.R
\name{first_not_na}
\alias{first_not_na}
\title{first_not_na}
\usage{
first_not_na(vec)
}
\arguments{
\item{vec}{A vector that may have initial NA values}
}
\description{
Return the first element of a vector that is not NA
}
