% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_deploy.R
\name{generate_xml_config}
\alias{generate_xml_config}
\title{generate_xml_config}
\usage{
generate_xml_config(job_type, github_https_address = NULL, email = NULL,
  args = NULL, job_name, job_code, project_url = NULL)
}
\description{
Utility to generate xml configs from template.
}
