% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_collating_schema.R
\name{get_lms_schemas}
\alias{get_lms_schemas}
\title{get_lms_schemas}
\usage{
get_lms_schemas(con)
}
\arguments{
\item{con}{Database connection from \code{mmkit::db_conn}}
}
\description{
List all known LMS schema names
}
\author{
Ilya Goldin
}
