% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uber_conf.R
\name{get_uber_conf}
\alias{get_uber_conf}
\title{get_uber_conf}
\usage{
get_uber_conf(access_token = Sys.getenv("GITHUB_PAT"), branch = "master",
  keep_prelaunch_programs = FALSE, force_update = TRUE)
}
\arguments{
\item{access_token}{Personal access token to github request.
Defaults to Sys.getenv('GITHUB_PAT')}

\item{branch}{Optional; branch of uber-conf repo from which to pull.
Default = master.}

\item{keep_prelaunch_programs}{Boolean value indicating whether unlaunched programs
should be returned in list of objects.  Default=FALSE.}

\item{force_update}{Force uberconf to get new data. Default = TRUE}
}
\description{
Collects and consolidates uber-conf program files from github.
}
