% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcols.R
\name{getcols}
\alias{getcols}
\title{getcols}
\usage{
getcols(table, long = FALSE, conn = NULL)
}
\arguments{
\item{table}{Character. The table to query (Schema optional)}

\item{long}{Logical. Should the data be displayed in long format? Defaults to FALSE}

\item{conn}{Connection object.}
}
\description{
SQL Convenience function. Gets column names and types from a specified table
}
