% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrs.R
\name{getrs}
\alias{getrs}
\title{getrs}
\usage{
getrs(statement, ..., conn = rs)
}
\arguments{
\item{statement}{SQL query string of data to return.}

\item{...}{Values to be passed as substitution params. See str_form for more detail.}

\item{conn}{A valid connection to a redshift database.}
}
\description{
A convenience function to execute database queries.
Supports str_form substitution params in statement by default.
}
