% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histvals.R
\name{histvals}
\alias{histvals}
\title{histvals}
\usage{
histvals(conn = rs, obj, field = NULL)
}
\arguments{
\item{conn}{A redshift database connection. Defaults to rs.}

\item{obj}{A staging table name. Schema is not required as salesforce_stg is assumed.}

\item{field}{History tracked field name.
Optional, if specified, returns unique values for field name listed.}
}
\description{
Return either a list of history tracked fields,
or specify a field and get unique values.
}
