% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_encode.R
\name{hot_encode}
\alias{hot_encode}
\title{hot_encode}
\usage{
hot_encode(input_col, output_label, blank_to_na = TRUE)
}
\arguments{
\item{input_col}{A character of factor vector to be coerced to a one-hot encoded
dataframe. Result should be bound to the original dataframe for modeling purposes.}

\item{output_label}{String used to label encoded columns. Eg: if variable to be
encoded is 'state', then the resulting column names will be: state_CA, state_MD,
state_AL, etc.}

\item{blank_to_na}{Converts blank strings to NA. Default = TRUE.}
}
\description{
A utility to create coerce a vector to a one-hot encoded dataframe.
}
