% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_conform_funs.R
\name{impute_gender}
\alias{impute_gender}
\title{impute_gender}
\usage{
impute_gender(names, fill.na = NA)
}
\arguments{
\item{names}{A list of first names against which to match.}

\item{fill.na}{Optionally replace missing values with string}
}
\description{
Imputes expected gender based on first names from social security
database.
}
