% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_github_tar.R
\name{install_github_tar}
\alias{install_github_tar}
\title{install_github_tar}
\usage{
install_github_tar(repo, access_token = Sys.getenv("GITHUB_PAT"),
  release = "latest", owner = "2uinc")
}
\arguments{
\item{repo}{Name of the repo that should be installed.}

\item{access_token}{A github pat or other valid access token to use if repo is private.
By default, checks for environment variable GITHUB_PAT.}

\item{release}{Version to install; Defaults to latest.}

\item{owner}{Owner of the repo; Defaults to 2uinc.}
}
\description{
Installs a repo from github by cloning the most recent release,
then installing from the local tarball.  This method avoids problems associated
with devtools::install_github when packages have private github sourced imports.
Primarily designed for use in deployment.
}
