% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_blank.R
\name{is_blank}
\alias{is_blank}
\title{is_blank}
\usage{
is_blank(string, blank_vals = c(NA, "NA", "na", ""))
}
\arguments{
\item{string}{A string to validate.}

\item{blank_vals}{A list of values to be considered blank.
Defaults to: c(NA, null, 'NA', 'na', '').}
}
\description{
Returns a boolean value indicating if a string is blank, or otherwise invalid.
By default, looks for any matches against: NA, null, 'NA', 'na', or ''.
}
