% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_yes_no.R
\name{is_yes_no}
\alias{is_yes_no}
\title{is_yes_no}
\usage{
is_yes_no(x)
}
\arguments{
\item{blank_vals}{A list of values to be considered blank.
Defaults to: c(NA, null, 'NA', 'na', '').}
}
\description{
Returns yes, no or NA.
}
