% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_deploy_template.R
\name{make_deploy_template}
\alias{make_deploy_template}
\title{make_deploy_template}
\usage{
make_deploy_template(execution_files = NULL,
  output_path = "jenkins/deploy.yaml", overwrite = FALSE)
}
\arguments{
\item{execution_files}{Paths to the files you want to execute in your job. Optional,
but must be added by hand later if not set.}

\item{output_path}{Path to deploy.yaml file that should be created.
Defaults to jenkins/deploy.yaml.}

\item{overwrite}{Remove existing file if found. Default = FALSE.}
}
\description{
Writes template file to deploy.yaml. Should be filled out before calling
deploy_project.
}
