% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_deploy.R
\name{prep_deploy_files}
\alias{prep_deploy_files}
\title{prep_deploy_files}
\usage{
prep_deploy_files(deploy_conf)
}
\arguments{
\item{deploy_conf}{locatation of deploy_conf.yaml file.}
}
\description{
Ensures proper directories exist, checks for fabfile conflicts,
and moves templates to correct locations.
}
