% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_collating_schema.R
\name{query_collating_schema}
\alias{query_collating_schema}
\title{query_collating_schema}
\usage{
query_collating_schema(con, query_template, schemas, n = Inf)
}
\arguments{
\item{con}{Database connection from \code{mmkit::db_conn}}

\item{query_template}{SQL query string templatized for \code{mmkit::str_form}}

\item{schemas}{Data frame with a column called \code{schema}}

\item{n}{Passed to \code{dplyr::collect}, in case your result set is large}
}
\description{
Runs the same query across multiple schemas in one DB and
collates results into one dataframe
}
\examples{
\dontrun{
con <- mmkit::db_conn(...)
s <- get_lms_schemas(con)
qu <- "SELECT foo FROM {{schema}}.foo_table"
query_collating_schema(con, qu, s)
}
}
\author{
Ilya Goldin
}
