% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\name{read_creds}
\alias{read_creds}
\title{read_creds}
\usage{
read_creds(database, ..., cred_location = "~/creds.json")
}
\description{
Read database credentials from a file like creds.json
See params in \code{\link{mmkit::db_conn}}
}
