% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_lst.R
\name{sql_lst}
\alias{sql_lst}
\title{sql_lst}
\usage{
sql_lst(X)
}
\arguments{
\item{X}{A vector or list}
}
\description{
A convenience funtion that combines a list or vector into a string \\n
which is then suitable for use in a sql statement
}
