% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_national.R
\docType{data}
\name{ssa_national}
\alias{ssa_national}
\title{ssa_national}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1603026 rows and 4 columns.}
\usage{
ssa_national
}
\description{
A dataframe containing gender classifications for common names,
sourced from the r-opensci/genderdata package.
}
\keyword{datasets}
