% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_form.R
\name{str_form}
\alias{str_form}
\title{str_form}
\usage{
str_form(file_or_string, ..., variable_identifier = c("{{", "}}"),
  default_char = "|", is_sql = FALSE)
}
\arguments{
\item{file_or_string}{A path to a file, or a string that includes substitution parameters}

\item{...}{An arbitrary number of named values corresponding to the substitution
params specified in the template. Can also be a named list.}

\item{variable_identifier}{Identifiers to denote paramters requiring substitution. Defaults to {{ and }}.}

\item{default_char}{default: pipe}

\item{is_sql}{If TRUE, str_form will collapse list arguments to string automatically.}
}
\description{
A poor mans version of python jinja2 templating
}
