% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UberConf.R
\docType{class}
\name{UberConf}
\alias{UberConf}
\alias{new}
\alias{get_all_pdo_names}
\alias{get_pdo}
\alias{get_all_value}
\alias{get_pdo_value}
\title{UberConf wrapper}
\format{\code{\link{R6Class}} object.}
\arguments{
\item{...}{Passed to \code{\link{get_uber_conf}}}

\item{pdo}{Program degree offering like \code{unc-mba}}

\item{key}{UberConf key like \code{salesforce_org}}

\item{path}{Path on disk, passed through \code{path.expand}}

\item{global}{Read UberConf data into an environment, so there's
a global copy. If FALSE, read data into a list, i.e., a copy that
only exists within one instance of this class.}

\item{reload}{Force a reload of a global copy}
}
\description{
OO API for reading UberConf

\code{new} Constructor

\code{get_all_pdo_name} Get all known program degree offering
names

\code{get_pdo} Get full UberConf tree for some program degree
offering. Does not compute any derived values.

\code{get_all_value} Get UberConf value across all program
degree offerings. Computes derived values as in \code{get_pdo_value}.

\code{get_pdo_value} Get UberConf value for a program degree
offering and specific key. Computes derived values for some known keys.
}
\examples{
\dontrun{
u <- UberConf$new() # Check out from GitHub
u <- UberConf$new(path="~/git/uber-conf/data") # Read from disk
u$get_pdo_value("unc-mba", "salesforce_org") # Key defined in UberConf
u$get_pdo_value("unc-mba", "segment_site") # Key yields derived value
}
}
