% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UberConfWriter.R
\docType{class}
\name{UberConfWriter}
\alias{UberConfWriter}
\alias{new}
\alias{get_all_pdo_names}
\alias{get_pdo}
\alias{set_all_key_value}
\alias{set_pdo_key_value}
\alias{write_json_all}
\alias{write_json_pdo}
\title{UberConf wrapper}
\format{\code{\link{R6Class}} object.}
\arguments{
\item{uber}{an object of class \code{UberConf}}

\item{pdo}{Program degree offering like \code{unc-mba}}

\item{key_value}{list like \code{list(sis=list(sis_bucket="au-mba"))}}

\item{path}{Path on disk, passed through \code{path.expand}}

\item{keep.null}{If false, removes entries that are set to NULL}
}
\description{
OO API for modifying and writing UberConf

\code{new} Constructor

\code{get_all_pdo_name} Get all known program degree offering
names

\code{get_pdo} Get full UberConf tree for some program degree
offering. Does not compute any derived values.

\code{set_all_key_value} Sets a constant value for all program
degree offerings.

\code{set_pdo_key_value} Sets a value for one program degree
offering.

\code{write_json_all} Writes JSON files for all program degree
offerings.

\code{write_json_pdo} Writes JSON file for one program degree
offering.
}
\examples{
\dontrun{
uw <- UberConfWriter$new(UberConf$new())
n <- list(sis = list(sis_bucket = NULL, university_id = NULL))
uw$set_all_key_value(n)
uw$write_json_all("mydir")
}
}
