% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric_deploy.R
\name{validate_deploy_entry}
\alias{validate_deploy_entry}
\title{validate_deploy_entry}
\usage{
validate_deploy_entry(job, file, args)
}
\arguments{
\item{job}{Name of job to be validated.}

\item{file}{file path to validate.}

\item{args}{Argument names to check in file}
}
\description{
Validates the .R function specified by the deployment file.
}
