% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_deploy_template.R
\name{validate_execution_files}
\alias{validate_execution_files}
\title{validate_execution_files}
\usage{
validate_execution_files(execution_files = NULL)
}
\description{
Checks that files specified for execution exist.
}
