% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_form.R
\name{vars_requested}
\alias{vars_requested}
\title{vars_requested}
\usage{
vars_requested(
  file_or_string,
  variable_identifier = c("{{", "}}"),
  default_char = "|"
)
}
\arguments{
\item{file_or_string}{A path to a file, or a string that includes substitution parameters}
}
\description{
Variables requested by str_form template
}
